feat(picker): storybook - expanded and improved chromatic template
Expands the Chromatic-only template in Storybook to increase coverage
of different options.

Organizes the stories into sections using the formatting and typography
previously established with Action button. States in each section are
based on the grid of options in the Desktop S2 design Figma, with some
additions such as the loading state. And the Switch examples from the
existing template.

Removes the "Open" story which is now covered in the Chromatic template.

Improves display of Popover within these different contexts; fixes how
it is aligned on the side label example. Changes the side label example
to use flex row instead of using inline-block which could wrap as the
screen shrunk.

Adds aria-labelledby and unique ids to make sure the label is associated
with the Picker.

Adds missing icon (to left of value/placeholder text) option to
Storybook, along with VRT coverage.